Adjust decoding of nullables and optionals
Fields that are typed `Optional` in electionguard-python are both
nullable and omittable. Fields that have (non-`None`) default values are
omittable but not nullable. `None` values are serialized as
`null`.

Note that the json schemas aren't fully reliable yet.

Notably, some proof and nonce fields are optional in the python code,
but setting them to optional here causes the compiler to yell, so I've
just marked those fields with comments in json.ts.